+++ src/service/graphql/typings.ts
@@ -13,6 +13,8 @@ export interface GraphQLContext {
     query?: HttpQueryRequest['query']
     graphqlResponse?: GraphQLResponse
     responseInit?: HttpQueryResponse['responseInit']
+    status?: 'success' | 'error'
+    graphqlErrors?: any[]
   }
 }