+++ src/service/http/index.ts
@@ -6,7 +6,7 @@ import { timer } from '../../utils/time'
 import { RouteHandler } from '../typings'
 import { clients } from './middlewares/clients'
 import { error } from './middlewares/error'
-import { logger } from './middlewares/logger'
+import { timings } from './middlewares/timings'
 
 export const createHttpRoute = <ClientsT extends IOClients, StateT, CustomT>(
   Clients: ClientsImplementation<ClientsT>,
@@ -14,6 +14,6 @@ export const createHttpRoute = <ClientsT extends IOClients, StateT, CustomT>(
 ) => {
   return (handler: RouteHandler<ClientsT, StateT, CustomT> | Array<RouteHandler<ClientsT, StateT, CustomT>>) => {
     const middlewares = Array.isArray(handler) ? handler : [handler]
-    return compose([clients(Clients, options), logger, error, ...middlewares].map(timer))
+    return compose([clients(Clients, options), timings, error, ...middlewares].map(timer))
   }
 }