+++ CHANGELOG.md
@@ -7,9 +7,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.8] - 2019-03-28
+
+## [3.0.8-beta] - 2019-03-28
+
+### Changed
+- Conflate 2xx and 5xx status labels to success and error respectively.
+- Adds `graphql-operation` metric that considers if _any_ resolver had errors, with two dimensions: success and error
+- Logs each resolver error individually and add request information
+- Stop logging successful route handlers
+- Add single hardcoded retry for sending error logs
+- Prepare for `graphql` route id (deprecating `__graphql`)
+- Disallow declaration of `graphql` as http route handler
+
+## [3.0.7-beta.0] - 2019-03-28
+
 ## [3.0.7] - 2019-03-28
 ### Added
 - tsErrorsAsWarnings parameter for the `linkApp` and `relinkApp` methods of the `Builder` class
+
+## [3.0.7-beta] - 2019-03-28
 
 ## [3.0.6] - 2019-03-28
 
@@ -18,6 +35,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [3.0.4] - 2019-03-27
 
 ## [3.0.3] - 2019-03-27
+
+### Added
+- Implement new Service() wrapper and port graphql route generation from service-runtime-node.
 
 ## [3.0.3-beta.4] - 2019-03-27